acceptance specs cleanup

+++ spec/acceptance/homepage_acceptance_spec.rb
@@ -2,13 +2,17 @@ require 'spec_helper'
 
 describe "homepage_spec integration" do
 
-  it "should be ok for good path" do
+  before do
     visit root_path
-    page.has_content?('powered by').must_equal true
-
-    LinkGroup.delete_all
-    visit root_path
-    page.has_content?('powered by').must_equal true
-
   end
+
+  it "should render fine" do
+    page.has_content?('powered by').must_equal true
+  end
+
+  if "should render fine with link_group deleted" do
+    LinkGroup.delete_all
+    page.has_content?('powered by').must_equal true
+  end
+
 end
